FEAT: 회원 정보 수정 구현 및 AWS S3 Config

+++ src/main/java/com/dnd/ground/domain/user/service/UserServiceImpl.java
@@ -20,6 +20,7 @@ import com.dnd.ground.domain.user.User;
 import com.dnd.ground.domain.user.dto.*;
 import com.dnd.ground.domain.user.repository.UserRepository;
 import com.dnd.ground.global.exception.CNotFoundException;
+import com.dnd.ground.global.exception.CNotValidationException;
 import com.dnd.ground.global.exception.CommonErrorCode;
 import com.dnd.ground.global.util.AmazonS3Service;
 import lombok.*;
@@ -30,7 +31,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -46,8 +46,8 @@ import static java.time.temporal.TemporalAdjusters.lastDayOfMonth;
  * @description 유저 서비스 클래스
  * @author  박세헌, 박찬호
  * @since   2022-08-01
- * @updated 1. 내 프로필 조회
- *          - 2022-10-13 박세헌
+ * @updated 1.회원 정보 수정 구현 완료
+ *          - 2022-10-22 박찬호
  */
 
 @Slf4j
@@ -383,35 +383,38 @@ public class UserServiceImpl implements UserService{
 
     /* 회원 프로필 수정 */
     @Transactional
-    public ResponseEntity<Boolean> editUserProfile(MultipartFile file, UserRequestDto.Profile requestDto){
-        String originalNick = requestDto.getOriginalNick();
-        String editNick = requestDto.getEditNick();
-        String intro = requestDto.getIntro();
+    public ResponseEntity<UserResponseDto.UInfo> editUserProfile(MultipartFile file, UserRequestDto.Profile requestDto){
+        String originalNick = requestDto.getOriginNickname();
+        String editNick = requestDto.getEditNickname();
+
+        //변경될 닉네임 중복 검사
+        if (!editNick.equals(originalNick) && !authService.validateNickname(editNick))
+            throw new CNotValidationException(CommonErrorCode.DUPLICATE_NICKNAME);
 
         User user = userRepository.findByNickname(originalNick).orElseThrow(
                 () -> new CNotFoundException(CommonErrorCode.NOT_FOUND_USER));
 
-        String pictureName;
-        String picturePath;
+        String intro = requestDto.getIntro();
+        String pictureName = user.getPictureName();
+        String picturePath = user.getPicturePath();
 
         // 기본 이미지로 변경
         if (requestDto.getIsBasic()){
             amazonS3Service.deleteFile(user.getPictureName());
             pictureName = "user/profile/default_profile.png";
             picturePath = "https://dnd-ground-bucket.s3.ap-northeast-2.amazonaws.com/user/profile/default_profile.png";
-            user.updatePicture(pictureName, picturePath);
         }
-        else{
-            // 기본 이미지가 아닌 유저의 사진으로 변경
+        else {
+            // 기본 이미지가 아닌 유저의 사진으로 변경 (프로필 사진 이름: 닉네임+카카오ID (Ex. NickA18345)
             if (!file.isEmpty()){
                 amazonS3Service.deleteFile(user.getPictureName());
-                Map<String, String> fileInfo = amazonS3Service.uploadToS3(file, "user/profile", createFileName(file.getOriginalFilename()));
+                Map<String, String> fileInfo = amazonS3Service.uploadToS3(file, "user/profile", editNick+user.getKakaoId().toString());
                 pictureName = fileInfo.get("fileName");
                 picturePath = fileInfo.get("filePath");
-                user.updatePicture(pictureName, picturePath);
             }
         }
-        user.updateProfile(editNick, intro);
+        user.updateProfile(editNick, intro, pictureName, picturePath);
+
         return authService.issuanceTokenByNickname(user.getNickname());
     }
 
@@ -436,17 +439,7 @@ public class UserServiceImpl implements UserService{
                 .collect(Collectors.toList()));
     }
 
-    private String createFileName(String fileName) {
-        return UUID.randomUUID().toString().concat(getFileExtension(fileName));
-    }
-
-    private String getFileExtension(String fileName) {
-        try {
-            return fileName.substring(fileName.lastIndexOf("."));
-        } catch (StringIndexOutOfBoundsException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "잘못된 형식의 파일(" + fileName + ") 입니다.");
-        }
-        
+    /*마이페이지 프로필 조회*/
     public UserResponseDto.Profile getUserProfile(String nickname){
         User user = userRepository.findByNickname(nickname).orElseThrow(
                 () -> new CNotFoundException(CommonErrorCode.NOT_FOUND_USER));
